requirements.txt
flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.2.1
pymongo==4.10.1
gunicorn==23.0.0
gevent==24.2.1
openai==2.16.0
Pillow==11.3.0
